+++ crates/test-ui/ui/mutability_traits_unimplementable2.rs
@@ -1,0 +1,8 @@
+//! Check that `mutability` traits are not implementable manually.
+use objc2::mutability;
+
+struct CustomStruct;
+
+unsafe impl mutability::IsMutable for CustomStruct {}
+
+fn main() {}